Header переименован в navbar и разбит на файлы
